pull args parsing to it's own mod
