+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+[v0.11.0] 2024-11-19
+-------------------
+
+- moved arguement parsting to module  
+
 [v0.10.0] 2024-11-19
 -------------------
